pyproject.toml
[project]
name = "flashpoint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "beautifulsoup4>=4.14.3",
    "celery>=5.6.2",
    "en-core-web-sm",
    "fastapi>=0.134.0",
    "feedparser>=6.0.12",
    "fpdf>=1.7.2",
    "google-genai>=1.0.0",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "lxml>=6.0.2",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.17.0",
    "redis>=7.3.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "spacy>=3.8.11",
    "sqlalchemy>=2.0.47",
    "telethon>=1.42.0",
    "typer>=0.24.1",
    "uvicorn>=0.41.0",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
